refactor: Remove unnecessary file, rename file- addPointerEventHandlers.ts 삭제. 더이상 사용하지 않습니다.- addEventListenerEffect의 파일 명을 add-event-listener-effect로 변경했습니다.
